pyproject.toml
[project]
name = "pyopengl-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.6",
    "pygame>=2.6.1",
    "pyopengl>=3.1.9",
]
